update models.py and gui.py files

+++ PycharmProjects/ToDoPy/models.py
@@ -9,35 +9,54 @@ def connect_db():
         return None
 
 def create_task(task_name):
-    try:
-        conn = connect_db()
+    conn = connect_db()
+    if conn is not None:
         cursor = conn.cursor()
         cursor.execute("INSERT INTO tasks (task_name, completed) VALUES (?, ?)", (task_name, 0))
         conn.commit()
-        print(f"Task '{task_name}' has been added to the database.")
-    except sqlite3.OperationalError as e:
-        print("Błąd bazy danych:", e)
-    finally:
         conn.close()
 
 def get_tasks():
     conn = connect_db()
-    cursor = conn.cursor()
-    cursor.execute("SELECT id, task_name, completed FROM tasks")
-    tasks = cursor.fetchall()
-    conn.close()
-    return tasks
+    if conn is not None:
+        cursor = conn.cursor()
+        cursor.execute("SELECT id, task_name, completed FROM tasks")
+        tasks = cursor.fetchall()
+        conn.close()
+        return tasks
+    return []
 
 def update_task(id, completed):
     conn = connect_db()
-    cursor = conn.cursor()
-    cursor.execute("UPDATE tasks SET completed = ? WHERE id = ?", (completed, id))
-    conn.commit()
-    conn.close()
+    if conn is not None:
+        cursor = conn.cursor()
+        cursor.execute("UPDATE tasks SET completed = ? WHERE id = ?", (completed, id))
+        conn.commit()
+        conn.close()
 
-def delete_task(task_id):
+def delete_task_from_db(task_id):
     conn = connect_db()
-    cursor = conn.cursor()
-    cursor.execute("DELETE FROM tasks WHERE id = ?", (task_id,))
-    conn.commit()
-    conn.close()
+    if conn is not None:
+        cursor = conn.cursor()
+        cursor.execute("DELETE FROM tasks WHERE id = ?", (task_id,))
+        conn.commit()
+        conn.close()
+
+def get_task_name(task_id):
+    conn = connect_db()
+    if conn is not None:
+        cursor = conn.cursor()
+        cursor.execute("SELECT task_name FROM tasks WHERE id = ?", (task_id,))
+        task_name = cursor.fetchone()
+        conn.close()
+        if task_name:
+            return task_name[0]
+    return None
+
+def update_task_name_in_db(task_id, new_name):
+    conn = connect_db()
+    if conn is not None:
+        cursor = conn.cursor()
+        cursor.execute("UPDATE tasks SET task_name = ? WHERE id = ?", (new_name, task_id))
+        conn.commit()
+        conn.close()
